services/recommender-ai-service/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-environ==0.11.2
psycopg2-binary==2.9.10
django-cors-headers==4.7.0
requests==2.32.3
gunicorn==23.0.0
whitenoise==6.8.2
